+++ src/main/java/com/example/demo/Service/StarService.java
@@ -4,8 +4,10 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 
+import com.example.demo.DTO.PlanetDTO;
 import com.example.demo.DTO.StarDTO;
 import com.example.demo.Entity.Star;
 import com.example.demo.Repository.StarRepository;
@@ -18,67 +20,71 @@ public class StarService {
 		this.starRepository = starRepository;
 	}
 	
-	public StarDTO getOne(int id) {
+	public StarDTO getOne(int id) throws Exception{
 		Optional<Star> tmpStar = starRepository.findById(id);
-		StarDTO starDTO = new StarDTO();
 		
 		try {
 			Star star = tmpStar.get();
-			starDTO.setId(star.getId());
-			starDTO.setName(star.getName());
-			starDTO.setDensity(star.getDensity());
-			return starDTO;
+			ModelMapper modelMapper = new ModelMapper();
+			return modelMapper.map(star, StarDTO.class);
 		} catch (Exception e) {
-			return starDTO;
+			throw new Exception();
 		}
 	}
 	
-	public List<StarDTO> getAll(){
+	public List<StarDTO> getAll() throws Exception{
 		List<StarDTO> starDTOs = new ArrayList<StarDTO>();
-		
-		for (Star star : starRepository.findAll() ) {
-			StarDTO starDTO = new StarDTO();
-			starDTO.setId(star.getId());
-			starDTO.setName(star.getName());
-			starDTO.setDensity(star.getDensity());
-			starDTOs.add(starDTO);
-		}
-		
-		return starDTOs;
-	}
-	
-	public StarDTO post(StarDTO starDTO) {
-		Star star = new Star();
-		star.setName(starDTO.getName());
-		star.setDensity(starDTO.getDensity());
+		ModelMapper modelMapper = new ModelMapper();
 		try {
-			starRepository.save(star);
+			for (Star star : starRepository.findAll() ) {
+				StarDTO starDTO = modelMapper.map(star, StarDTO.class);
+				starDTOs.add(starDTO);
+			}
+			
+			return starDTOs;
 		} catch (Exception e) {
-			// TODO: handle exception
+			throw new Exception();
+		}
+	}
+	
+	public StarDTO post(StarDTO starDTO) throws Exception{
+		ModelMapper modelMapper = new ModelMapper();
+		try {
+			Star star = modelMapper.map(starDTO, Star.class);
+			//Por standar se devuelve un objeto
+			star = starRepository.save(star);
+			return modelMapper.map(star, StarDTO.class);
+		} catch (Exception e) {
+			throw new Exception();
 		}
 		
-		starDTO.setId(star.getId());
-		return starDTO;
 	}
 	
-	public StarDTO put(StarDTO starDTO, int id) {
-		Optional<Star> temp = starRepository.findById(id);			
-		try {		
+	public StarDTO put(StarDTO starDTO, int id) throws Exception {
+		Optional<Star> temp = starRepository.findById(id);		
+		ModelMapper modelMapper = new ModelMapper();
+
+		try {	
 			Star star = temp.get();
-			star.setName(starDTO.getName());
-			star.setDensity(starDTO.getDensity());
-			starRepository.save(star);		
+			star = modelMapper.map(starDTO, Star.class);
+			star = starRepository.save(star);		
 			starDTO.setId(star.getId());		
+			return modelMapper.map(star, StarDTO.class);			 
 		} catch (Exception e) {			
+			throw new Exception();
 		}			
-		return starDTO;	
 	}
 	
-	public boolean delete(int id) {
-		Optional<Star> temp = starRepository.findById(id);	
-		try {					
-			starRepository.delete(temp.get());
-			return true;		
+	public boolean delete(int id) throws Exception {
+		try {				
+			if(starRepository.existsById(id)) {
+				Optional<Star> temp = starRepository.findById(id);	
+				starRepository.delete(temp.get());
+				return true;
+			}else {
+				return false;
+			}
+					
 		} catch (Exception e) {
 			return false;
